Made Number properties bindable to each other

+++ planet-examples/src/main/java/org/saar/example/Check.java
@@ -1,17 +1,18 @@
 package org.saar.example;
 
 import org.saar.utils.property.type.FloatProperty;
+import org.saar.utils.property.type.IntProperty;
 
 public class Check {
 
     public static void main(String[] args) {
         final FloatProperty property1 = new FloatProperty(4f);
-        final FloatProperty property2 = new FloatProperty(8f);
+        final IntProperty property2 = new IntProperty(8);
 
         property1.bindBidirectional(property2);
 
         System.out.println(property1.get());
-        property2.set(7f);
+        property2.set(7);
         System.out.println(property1.get());
         property1.set(3f);
         System.out.println(property1.get());
